requirements.txt
torch>=1.9.0
torchvision>=0.10.0
numpy>=1.19.2
Pillow>=8.3.1
tqdm>=4.62.3
wandb>=0.12.0
pyyaml>=5.4.1
matplotlib>=3.4.3 